keyboardMove refactor wip
